server_files/requirements.txt
Flask==2.0.*
pymongo==3.7.2
PyYAML==6.0.*